Fixing api issues
modified redFlagController
modified userController
modified redFlagValidation
added validateRedFlag
modified redFlagModel
modified userModel
modified redFlagRouter
added authTest
modified test
modified packageJson

[finishes #169998939]
