+++ Server/helper/redFlagValidation.js
@@ -3,11 +3,11 @@ import Joi from '@hapi/joi';
 const redFlagSchema = Joi.object().keys({
   id: Joi.number(),
   createdOn: Joi.required(),
-  title: Joi.string().regex(/^[a-zA-Z]{3,20}$/).required(),
-  type: Joi.string().regex(/^[a-zA-Z]{3,20}$/).required(),
+  title: Joi.string().trim().regex(/^[a-zA-Z]{3,20}$/).required(),
+  type: Joi.string().trim().regex(/^[a-zA-Z]{3,20}$/).required(),
   comment: Joi.string().trim().required(),
-  location: Joi.string().required(),
-  status: Joi.string().regex(/^[a-zA-Z]{3,20}$/).required(),
+  location: Joi.string().trim().required(),
+  status: Joi.string().trim().regex(/^[a-zA-Z]{3,20}$/).required(),
   userId: Joi.number(),
 });
