+++ Server/model/redFlagModel.js
@@ -1,44 +1,3 @@
-const redFlag = [
-  {
-    id: 1,
-    createdOn: 'November 18, 2019 10:15 AM',
-    title: 'corruption',
-    type: 'redFlag',
-    comment: 'this corruption was very dangerous',
-    location: '-1.978400, 30.226990',
-    status: 'solved',
-    userId: 1,
-  },
-  {
-    id: 2,
-    createdOn: 'November 10, 2019 2:30 PM',
-    title: 'collupsedbridge',
-    type: 'intervention',
-    comment: 'this bridge has being collupsed completely',
-    location: '56.349510, 15.928380',
-    status: 'rejected',
-    userId: 2,
-  },
-  {
-    id: 3,
-    createdOn: 'October 18, 2019 4:03 PM',
-    title: 'collupsedroad',
-    type: 'intervention',
-    comment: 'this road has being collupsed completely',
-    location: '-2.8935, 327849',
-    status: 'draft',
-    userId: 5,
-  },
-  {
-    id: 4,
-    createdOn: 'Jannuary 18, 2019 1:03 PM',
-    title: 'collupsedroad',
-    type: 'intervention',
-    comment: 'this road has being collupsed completely',
-    location: '-1.8935, 307849',
-    status: 'draft',
-    userId: 4,
-  },
-];
+const redFlag = [];
 
 export default redFlag;
