+++ Server/test/test.js
@@ -8,110 +8,7 @@ chai.should();
 const userToken = generateToken.generate('andrea@gmail.com', 'user');
 describe('BroadCaster Api', () => {
   const wrongToken = 'IsInR5cCI6IkpXVCJ9.eyJpZCI6NSwiZW1haWwiOiJ0YUBnbWFpbC5jb20iLCJ1c2VyVHlwZSI6InVzZXIiLCJpYXQiOjE1NzM0NzgzNzZ9.9rJCaBlriG8j5JDqRTkhB-SICNc';
-  // signUp test
-  it('should be signUp', (done) => {
-    const user = {
-      firstName: 'kambanda',
-      lastName: 'andrea',
-      email: 'andrea@gmail.com',
-      phoneNumber: '0786960926',
-      userName: 'andrea',
-      password: 'andrea123',
-      userType: 'user',
-    };
-    chai.request(server)
-      .post('/api/v1/signUp')
-      .send(user)
-      .end((error, res) => {
-        res.status.should.be.equal(201);
-        res.body.message.should.be.equal('user created successfully');
-      });
-    done();
-  });
-  // duplication email test
-  it('should not be signup with email already exist.', (done) => {
-    const user = {
-      firstName: 'ruhimbaza',
-      lastName: 'bertin',
-      email: 'andrea@gmail.com',
-      phoneNumber: '0786960926',
-      userName: 'andrea',
-      password: 'andrea123',
-      userType: 'user',
-    };
-    chai.request(server)
-      .post('/api/v1/signUp')
-      .send(user)
-      .end((error, res) => {
-        res.status.should.be.equal(409);
-        res.body.error.should.be.equal('email already exist, Please Login');
-      });
-    done();
-  });
-  // user should not be able to signup if some fields are not filled yet
-  it('should not be able to signup when some filled are missing', (done) => {
-    const user = {
-      lastName: 'bertin',
-      email: 'andrea@gmail.com',
-      phoneNumber: '0786960926',
-      userName: 'andrea',
-      password: 'andrea123',
-      userType: 'user',
-    };
-    chai.request(server)
-      .post('/api/v1/signUp')
-      .send(user)
-      .end((error, res) => {
-        res.status.should.be.equal(400);
-      });
-    done();
 
-  });
-  // signin test
-  it('should login', (done) => {
-    const user = {
-      email: 'andrea@gmail.com',
-      password: 'andrea123',
-    };
-    chai.request(server)
-      .post('/api/v1/signIn')
-      .send(user)
-      .end((error, res) => {
-        res.status.should.be.equal(200);
-        res.body.message.should.be.equal('Logged in as kambanda');
-      });
-    done();
-  });
-  // user should not be able to login if email or password is not correct.
-  it('should not be able to login if userName is not correct', (done) => {
-    const user = {
-      email: 'andre@gmail.com',
-      password: 'jeanete123',
-    };
-    chai.request(server)
-      .post('/api/v1/signIn')
-      .send(user)
-      .end((error, res) => {
-        res.status.should.be.equal(401);
-        res.body.error.should.be.equal('Incorrect UserName or Password');
-      });
-    done();
-  });
-  // user should not be able to login if password is not correct.
-  it('should not be able to login if userName is not correct', (done) => {
-    const user = {
-      userName: 'andre@gmail.com',
-      password: 'bertin123',
-    };
-    chai.request(server)
-      .post('/api/v1/signIn')
-      .send(user)
-      .end((error, res) => {
-        res.status.should.be.equal(401);
-        res.body.error.should.be.equal('Incorrect UserName or Password');
-      });
-    done();
-  });
   // Create a new RedFlag
   it('should create a new RedFlag record', (done) => {
     const newRedFlag = {
@@ -119,7 +16,7 @@ describe('BroadCaster Api', () => {
       type: 'redFlag',
       comment: 'this corruption was very dangerous',
       location: 'Gasabo',
-      status: 'solved',
+      status: 'draft',
     };
     chai.request(server)
       .post('/api/v1/redFlags')
@@ -138,15 +35,15 @@ describe('BroadCaster Api', () => {
       type: 'redFlag',
       comment: 'this corruption was very dangerous',
       location: 'Gasabo',
-      status: 'solved',
+      status: 'draft',
     };
     chai.request(server)
       .post('/api/v1/redFlags')
       .set('token', wrongToken)
       .send(newRedFlag)
       .end((error, res) => {
-        res.status.should.be.equal(403);
-        res.body.error.should.be.equal('Authentication failed');
+        res.status.should.be.equal(401);
+        res.body.error.should.be.equal('Invalid token');
       });
     done();
   });
@@ -182,15 +79,15 @@ describe('BroadCaster Api', () => {
       .get('/api/v1/red-flags')
       .set('token', wrongToken)
       .end((error, res) => {
-        res.status.should.be.equal(403);
-        res.body.error.should.be.equal('Authentication failed');
+        res.status.should.be.equal(401);
+        res.body.error.should.be.equal('Invalid token');
       });
     done();
   });
   // view a specific redFlag created
   it('should be able to view a specific redFlag created', (done) => {
     chai.request(server)
-      .get('/api/v1/red-flags/2')
+      .get('/api/v1/red-flags/1')
       .set('token', userToken)
       .end((error, res) => {
         res.status.should.be.equal(200);
@@ -203,8 +100,8 @@ describe('BroadCaster Api', () => {
       .get('/api/v1/red-flags/2')
       .set('token', wrongToken)
       .end((error, res) => {
-        res.status.should.be.equal(403);
-        res.body.error.should.be.equal('Authentication failed');
+        res.status.should.be.equal(401);
+        res.body.error.should.be.equal('Invalid token');
       });
     done();
   });
@@ -248,7 +145,7 @@ describe('BroadCaster Api', () => {
       status: 'draft',
     };
     chai.request(server)
-      .patch('/api/v1/red-flags/3/comment')
+      .patch('/api/v1/red-flags/1/comment')
       .set('token', userToken)
       .send(newRedFlag)
       .end((error, res) => {
@@ -263,7 +160,7 @@ describe('BroadCaster Api', () => {
       type: 'redFlag',
       comment: 'This is a huge disaster',
       location: 'Gasabo',
-      status: 'solved',
+      status: 'draft',
     };
     chai.request(server)
       .patch('/api/v1/red-flags/35/comment')
@@ -284,7 +181,7 @@ describe('BroadCaster Api', () => {
       status: 'draft',
     };
     chai.request(server)
-      .patch('/api/v1/red-flags/3/location')
+      .patch('/api/v1/red-flags/1/location')
       .set('token', userToken)
       .send(newRedFlag)
       .end((error, res) => {
@@ -317,7 +214,7 @@ describe('BroadCaster Api', () => {
       location: '-1.8935, 30.27849',
     };
     chai.request(server)
-      .patch('/api/v1/red-flags/3/location')
+      .patch('/api/v1/red-flags/1/location')
       .set('token', userToken)
       .send(newRedFlag)
       .end((error, res) => {
@@ -328,7 +225,7 @@ describe('BroadCaster Api', () => {
   // Only the user who created the ​red-flag is the one to delete the record.
   it('should not be able to delete the redFlag which is not created by you', (done) => {
     chai.request(server)
-      .delete('/api/v1/red-flags/3')
+      .delete('/api/v1/red-flags/1')
       .set('token', userToken)
       .end((error, res) => {
         res.status.should.be.equal(403);
@@ -354,7 +251,7 @@ describe('BroadCaster Api', () => {
         res.body.message.should.be.equal('RedFlag created successifully');
       });
     chai.request(server)
-      .delete('/api/v1/red-flags/6')
+      .delete('/api/v1/red-flags/2')
       .set('token', userToken)
       .end((error, res) => {
         res.status.should.be.equal(200);
@@ -386,10 +283,10 @@ describe('BroadCaster Api', () => {
   //  User should not be able to delete the record if that is not the one who created it.
   it('should not be able to delete the record if you are not the one to create it', (done) => {
     chai.request(server)
-      .delete('/api/v1/red-flags/1')
+      .delete('/api/v1/red-flags/2')
       .set('token', userToken)
       .send((error, res) => {
-        res.status.should.be.equal(403);
+        res.status.should.be.equal(401);
       });
     done();
   });
