+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon --exec babel-node server/app.js",
+    "start": "nodemon --exec babel-node server/app.js",
     "test": "nyc --reporter=html --reporter=text mocha --require babel-register --timeout 10000 --exit  Server/test/*",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
